chore(apps/web): refactor component to add search UI
