+++ apps/web/src/components/inputs/inputs.tsx
@@ -2,10 +2,11 @@
 
 import { Stack } from "@mantine/core";
 import { FC, useCallback, useState } from "react";
-import { InputOrderByInput } from "../../graphql/explorer/types";
-import { useInputsQuery } from "../../graphql/explorer/hooks/queries";
+import { InputOrderByInput, useInputsConnectionQuery } from "../../graphql";
+import { checkQuery } from "../../lib/query";
 import InputsTable from "../inputs/inputsTable";
 import Paginated from "../paginated";
+import Search from "../search";
 
 export type InputsProps = {
     orderBy?: InputOrderByInput;
@@ -16,15 +17,20 @@ const Inputs: FC<InputsProps> = ({
     orderBy = InputOrderByInput.TimestampDesc,
     applicationId,
 }) => {
+    const [query, setQuery] = useState("");
     const [limit, setLimit] = useState(10);
     const [page, setPage] = useState(1);
     const after = page === 1 ? undefined : ((page - 1) * limit).toString();
-    const [{ data, fetching }] = useInputsQuery({
+
+    const [{ data: data, fetching: fetching }] = useInputsConnectionQuery({
         variables: {
             orderBy,
-            applicationId: applicationId?.toLowerCase(),
             limit,
             after,
+            where: checkQuery(
+                query.toLowerCase(),
+                applicationId?.toLowerCase(),
+            ),
         },
     });
     const inputs = data?.inputsConnection.edges.map((edge) => edge.node) ?? [];
@@ -34,8 +40,13 @@ const Inputs: FC<InputsProps> = ({
         setPage(page);
     }, []);
 
+    const onSubmitQuery = useCallback((query: string) => {
+        setQuery(query);
+    }, []);
+
     return (
         <Stack>
+            <Search onSubmit={onSubmitQuery} />
             <Paginated
                 fetching={fetching}
                 totalCount={data?.inputsConnection.totalCount}
